requirements.txt
langchain
streamlit
openai
python-decouple
langchain_community
langchain_openai
rootpath
pypdf
faiss-cpu
python-dotenv
pymupdf